vote-app/requirements.txt
Flask==2.3.2
Werkzeug==2.3.4
redis==3.5.3
psycopg2-binary==2.9.3  # Cambiado a psycopg2-binary
pandas==1.3.3
requests==2.28.1  # Añadir la dependencia de requests
